refactoring completed

+++ src/main/java/com/etiya/rentACarSpring/businnes/concretes/BrandManager.java
@@ -2,14 +2,12 @@ package com.etiya.rentACarSpring.businnes.concretes;
 
 import com.etiya.rentACarSpring.businnes.abstracts.MessageService;
 import com.etiya.rentACarSpring.businnes.dtos.BrandSearchListDto;
-import com.etiya.rentACarSpring.businnes.dtos.RentalSearchListDto;
 import com.etiya.rentACarSpring.core.utilities.results.*;
-import com.etiya.rentACarSpring.entities.Rental;
+
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import com.etiya.rentACarSpring.businnes.abstracts.BrandService;
-import com.etiya.rentACarSpring.businnes.constants.Messages;
 import com.etiya.rentACarSpring.businnes.request.BrandRequest.CreateBrandRequest;
 import com.etiya.rentACarSpring.businnes.request.BrandRequest.DeleteBrandRequest;
 import com.etiya.rentACarSpring.businnes.request.BrandRequest.UpdateBrandRequest;
@@ -48,7 +46,8 @@ public class BrandManager implements BrandService {
 
     @Override
     public Result save(CreateBrandRequest createBrandRequest) {
-        Result result = BusinnessRules.run(checkBrandNameDublicated(createBrandRequest.getBrandName()));
+        Result result = BusinnessRules.run(checkBrandNameDublicated(createBrandRequest.getBrandName())
+               );
         if (result != null) {
             return result;
         }
@@ -61,7 +60,8 @@ public class BrandManager implements BrandService {
 
     @Override
     public Result update(UpdateBrandRequest updateBrandRequest) {
-        Result result = BusinnessRules.run(checkBrandNameDublicated(updateBrandRequest.getBrandName()));
+        Result result = BusinnessRules.run(checkBrandNameDublicated(updateBrandRequest.getBrandName()),
+                checkIfBrandExists(updateBrandRequest.getBrandId()));
         if (result != null) {
             return result;
         }
@@ -73,9 +73,21 @@ public class BrandManager implements BrandService {
 
     @Override
     public Result delete(DeleteBrandRequest deleteBrandRequest) {
+        Result result = BusinnessRules.run(checkIfBrandExists(deleteBrandRequest.getBrandId()));
+        if (result != null) {
+            return result;
+        }
 
         this.brandDao.deleteById(deleteBrandRequest.getBrandId());
         return new SuccesResult(messageService.getByEnglishMessageByMessageId(3));
+    }
+
+    @Override
+    public Result checkIfBrandExists(int brandId) {
+        if (!this.brandDao.existsById(brandId)) {
+            return new ErrorResult("brand Id bulunamadı");
+        }
+        return new SuccesResult();
     }
 
     private Result checkBrandNameDublicated(String brandName) {
@@ -86,5 +98,7 @@ public class BrandManager implements BrandService {
 
         return new SuccesResult();
     }
+
+
 
 }
